refactor: resolved conflict, moved code to external files
