+++ @uportal/content-carousel/src/App.vue
@@ -1,18 +1,5 @@
 <template>
   <ContentCarousel :strategy="demoStrategy"/>
 </template>
-
-<script lang="ts">
-import { Component, Provide, Vue } from 'vue-property-decorator';
-import ContentCarousel from './components/ContentCarousel.vue';
-import { demoStrategy } from './lib/DemoStrategy';
-
-@Component({
-  components: {
-    ContentCarousel,
-  },
-})
-export default class App extends Vue {
-  @Provide() private demoStrategy = demoStrategy;
-}
+<script lang="ts" src="./App.ts">
 </script>
