+++ @uportal/content-carousel/src/main.ts
@@ -1,5 +1,5 @@
 import Vue from 'vue';
-import App from './App.vue';
+import App from './App';
 
 Vue.config.productionTip = false;
